% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer-core.R
\name{create_results}
\alias{create_results}
\title{Create results object from batch}
\usage{
create_results(result)
}
\arguments{
\item{result}{Batch object}
}
\value{
Results object with class "batch"
}
\description{
Create results object from batch
}
\keyword{internal}
