% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer-core.R
\name{handle_batch_interrupt}
\alias{handle_batch_interrupt}
\title{Handle batch interruption}
\usage{
handle_batch_interrupt(result, beep)
}
\arguments{
\item{result}{A batch object containing processing state}

\item{beep}{Logical indicating whether to play a sound}
}
\value{
NULL (called for side effects)
}
\description{
Handle batch interruption
}
\keyword{internal}
