% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{package}
\name{hellmer-package}
\alias{hellmer}
\alias{hellmer-package}
\title{hellmer: Batch Processing for Chat Models}
\description{
Batch processing framework for 'ellmer' chat models. Provides both sequential and parallel processing of chat interactions with features including tool calling and structured data extraction. Enables workflow management through progress tracking and recovery and automatic retry with backoff. Additional quality-of-life features include verbosity (or echo) control and sound notifications. Parallel processing is implemented via the 'future' framework. Includes methods for retrieving progress status, chat texts, and chat objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dylanpieper.github.io/hellmer/}
}

}
\author{
\strong{Maintainer}: Dylan Pieper \email{dylanpieper@gmail.com}

}
\keyword{internal}
