% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hellmer-core.R
\name{process_judgements}
\alias{process_judgements}
\title{Process structured data extraction with judgement}
\usage{
process_judgements(
  chat_obj,
  prompt,
  type_spec,
  judgements = 0,
  echo = FALSE,
  ...
)
}
\arguments{
\item{chat_obj}{Chat model object}

\item{prompt}{The prompt or text to analyze}

\item{type_spec}{Type specification for structured data}

\item{judgements}{Number of judgements for structured data extraction resulting in refined data}
}
\value{
List containing extraction process
}
\description{
Process structured data extraction with judgement
}
\keyword{internal}
